Fixed bug where parameters were not being loaded into the cat_creator node. Changed CAT calculation to use above-baseline channels only

+++ arm/src/arm_control.cpp
@@ -125,7 +125,7 @@ void ArmControl::constantMoveCallback(const arm::constant_move::ConstPtr& cmd)
         else
         {
             for (int i = 0; i < STATE_ARR_SZ; i++)
-                states_[i] = cmd->states[i];
+                states_[i] = static_cast<int>(cmd->states[i]);
             arm_->moveConstant(states_);
         }
     }
